pyproject.toml
[project]
name = "elastic_semantic_search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch==8.13.0",
    "fastapi>=0.116.1",
    "jsonschema>=4.25.0",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "uvicorn[standard]>=0.35.0",
]
